backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.11",
    "google-generativeai>=0.8.4",
    "pinecone[grpc]>=6.0.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=3.4.1",
    "uvicorn>=0.34.0",
]
